feat: refactor example queries to single form with dimensions component (#317)
