+++ app/views/ExampleQueriesView/components/Dimensions/components/DimensionSection/components/FacetValueList/facetValueList.styles.ts
@@ -1,13 +1,5 @@
 import { bpDownSm } from "@databiosphere/findable-ui/lib/styles/common/mixins/breakpoints";
 import styled from "@emotion/styled";
-import { Stack } from "@mui/material";
-
-export const StyledChipStack = styled(Stack)`
-  flex-direction: row;
-  flex-wrap: wrap;
-  gap: 8px;
-  margin: 16px 0;
-`;
 
 export const StyledValueList = styled("ul")`
   column-count: 3;
