+++ app/views/ExampleQueriesView/components/Dimensions/components/DimensionSection/dimensionSection.styles.ts
@@ -1,0 +1,9 @@
+import { Chips } from "@databiosphere/findable-ui/lib/views/ResearchView/assistant/components/Messages/components/PromptMessage/components/Chips/chips";
+import styled from "@emotion/styled";
+
+export const StyledChips = styled(Chips)`
+  flex-direction: row;
+  flex-wrap: wrap;
+  gap: 8px;
+  margin: 16px 0;
+`;
