+++ app/views/ExampleQueriesView/components/Dimensions/components/DimensionSection/types.ts
@@ -1,0 +1,3 @@
+import { Dimension } from "../../constants";
+
+export type DimensionSectionProps = Dimension;
