+++ app/views/ExampleQueriesView/components/Dimensions/dimensions.styles.ts
@@ -1,0 +1,5 @@
+import styled from "@emotion/styled";
+
+export const StyledForm = styled("form")`
+  min-width: 0;
+`;
