+++ app/views/ExampleQueriesView/exampleQueriesView.tsx
@@ -2,8 +2,7 @@ import { LAYOUT_STYLE_NO_CONTRAST_DEFAULT } from "@databiosphere/findable-ui/lib
 import { ContentView } from "@databiosphere/findable-ui/lib/views/ContentView/contentView";
 import { JSX } from "react";
 import { Content } from "../../components/Layout/components/Content/content";
-import { DimensionSection } from "./components/DimensionSection/dimensionSection";
-import { DIMENSIONS } from "./constants";
+import { Dimensions } from "./components/Dimensions/dimensions";
 
 /**
  * Renders the example queries page with interactive query chips organized by search dimension.
@@ -21,9 +20,7 @@ export const ExampleQueriesView = (): JSX.Element => {
             natural language — the search engine maps your words to the
             appropriate catalog facets automatically.
           </p>
-          {DIMENSIONS.map((dimension) => (
-            <DimensionSection key={dimension.title} dimension={dimension} />
-          ))}
+          <Dimensions />
         </Content>
       }
     />
